Moved subgrid functionality to entity
