+++ Microsoft.Dynamics365.UIAutomation.Api/Pages/Grid.cs
@@ -426,142 +426,5 @@ namespace Microsoft.Dynamics365.UIAutomation.Api
             });
         }
 
-        /// <summary>
-        /// Retrieve the items from a subgrid.
-        /// </summary>
-        /// <param name="subgridName">Unique name (not label) of the subgrid.</param>
-        /// <param name="thinkTime">Used to simulate a wait time between human interactions. The Default is 2 seconds.</param>
-        /// <example>xrmBrowser.Grid.GetSubGridItems("CONTACTS");</example>
-        public BrowserCommandResult<List<GridItem>> GetSubGridItems(string subgridName, int thinkTime = Constants.DefaultThinkTime)
-        {
-            Browser.ThinkTime(thinkTime);
-
-            return this.Execute(GetOptions($"Get Subgrid Items for Subgrid {subgridName}"), driver =>
-            {
-                List<GridItem> subGridRows = new List<GridItem>();
-
-                if (!driver.HasElement(By.XPath(Elements.Xpath[Reference.Entity.SubGrid].Replace("[NAME]", subgridName))))
-                    throw new NotFoundException($"{subgridName} subgrid not found. Subgrid names are case sensitive.  Please make sure casing is the same.");
-
-                //Find the subgrid contents
-                var subGrid = driver.WaitUntilAvailable(By.XPath(Elements.Xpath[Reference.Entity.SubGrid].Replace("[NAME]", subgridName)));
-
-                //Find the columns
-                List<string> columns = new List<string>();
-
-                var headerCells = subGrid.FindElements(By.XPath(Elements.Xpath[Reference.Entity.SubGridHeaderCell].Replace("[NAME]", subgridName)));
-
-                foreach (IWebElement headerCell in headerCells)
-                {
-                    var headerText = headerCell.GetAttribute("innerText");
-                    if (!string.IsNullOrEmpty(headerText))
-                        columns.Add(headerText);
-                }
-
-                //Find the rows
-                var rows = subGrid.FindElements(By.XPath(Elements.Xpath[Reference.Entity.SubGridRow].Replace("[NAME]", subgridName)));
-
-                //Process each row
-                foreach (IWebElement row in rows)
-                {
-                    List<string> cellValues = new List<string>();
-                    GridItem item = new GridItem
-                    {
-                        //Get the entityId and entity Type
-                        Id = Guid.Parse(row.GetAttribute("oid")),
-                        EntityName = row.GetAttribute("otypename")
-                    };
-
-                    var cells = row.FindElements(By.XPath(Elements.Xpath[Reference.Entity.SubGridCell].Replace("[NAME]", subgridName)));
-
-                    if (cells.Count > 0)
-                    {
-                        foreach (IWebElement thisCell in cells)
-                            cellValues.Add(thisCell.Text);
-
-                        for (int i = 0; i < columns.Count; i++)
-                        {
-                            //The first cell is always a checkbox for the record.  Ignore the checkbox.
-                            item[columns[i]] = cellValues[i + 1];
-                        }
-
-                        subGridRows.Add(item);
-                    }
-                }
-
-                return subGridRows;
-            });
-        }
-
-        /// <summary>
-        /// Retrieves the number of rows from a subgrid
-        /// </summary>
-        /// <param name="subgridName">Unique name (not label) of the subgrid to retrieve items from</param>
-        /// <example>xrmBrowser.Grid.GetSubGridItemsCount("CONTACTS");</example>
-        public BrowserCommandResult<int> GetSubGridItemsCount(string subgridName)
-        {
-            return this.Execute(GetOptions($"Get Subgrid Items Count for subgrid { subgridName}"), driver =>
-            {
-                List<GridItem> rows = GetSubGridItems(subgridName);
-                return rows.Count;
-            });
-        }
-
-        /// <summary>
-        /// Open a record on a subgrid.
-        /// </summary>
-        /// <param name="subgridName">Unique name (not label) of the subgrid.</param>
-        /// <param name="index">Index of the record to open.</param>
-        /// <param name="thinkTime">Used to simulate a wait time between human interactions. The Default is 2 seconds.</param>
-        /// <example>xrmBrowser.Grid.OpenSubGridRecord("CONTACTS", 0);</example>
-        public BrowserCommandResult<bool> OpenSubGridRecord(string subgridName, int index = 0, int thinkTime = Constants.DefaultThinkTime)
-        {
-            Browser.ThinkTime(thinkTime);
-
-            return this.Execute(GetOptions($"Open Subgrid record for subgrid { subgridName }"), driver =>
-            {
-                var subGrid = driver.WaitUntilAvailable(By.XPath(Elements.Xpath[Reference.Entity.SubGrid].Replace("[NAME]", subgridName)));
-
-                var subGridRows = subGrid.FindElements(By.XPath(Elements.Xpath[Reference.Entity.SubGridRow].Replace("[NAME]", subgridName)));
-                if (index + 1 > subGridRows.Count)
-                    throw new InvalidOperationException($"No record with the index '{index}' exists.");
-
-                driver.DoubleClick(subGridRows[index]);
-
-                return true;
-            });
-        }
-
-        /// <summary>
-        /// Deletes a record from a subgrid.
-        /// </summary>
-        /// <param name="subgridName">Unique name (not label) of the subgrid.</param>
-        /// <param name="index">Index of the record to delete.</param>
-        /// <param name="thinkTime">Used to simulate a wait time between human interactions. The Default is 2 seconds.</param>
-        /// <example>xrmBrowser.Grid.DeleteSubGridRecord("CONTACTS", 0);</example>
-        public BrowserCommandResult<bool> DeleteSubGridRecord(string subgridName, int index = 0, int thinkTime = Constants.DefaultThinkTime)
-        {
-            Browser.ThinkTime(thinkTime);
-
-            return this.Execute(GetOptions($"Delete Subgrid record for subgrid { subgridName }"), driver =>
-            {
-                var subGrid = driver.WaitUntilAvailable(By.XPath(Elements.Xpath[Reference.Entity.SubGrid].Replace("[NAME]", subgridName)));
-
-                var subGridRows = subGrid.FindElements(By.XPath(Elements.Xpath[Reference.Entity.SubGridRow].Replace("[NAME]", subgridName)));
-                if (index + 1 > subGridRows.Count)
-                    throw new InvalidOperationException($"No record with the index '{index}' exists.");
-
-                subGridRows[index].Hover(driver);
-
-                var recordId = subGridRows[index].GetAttribute("oid");
-                driver.ClickWhenAvailable(By.XPath(Elements.Xpath[Reference.Entity.SubGridRowDeleteButton].Replace("[RECORDID]", recordId)));
-
-                Dialog deleteDialog = new Dialog(Browser);
-                deleteDialog.Delete();
-
-                return true;
-            });
-        }
-
     }
 }
