+++ Microsoft.Dynamics365.UIAutomation.Sample/Web/Entity/SubGrid.cs
@@ -33,7 +33,7 @@ namespace Microsoft.Dynamics365.UIAutomation.Sample.Web
                 xrmBrowser.Grid.OpenRecord(0);
 
                 //This uses the unique name (not label) from the subgrid properties
-                xrmBrowser.Grid.OpenSubGridRecord("CONTACTS", 0);
+                xrmBrowser.Entity.OpenSubGridRecord("CONTACTS", 0);
 
                 xrmBrowser.ThinkTime(2000);
             }
@@ -55,7 +55,7 @@ namespace Microsoft.Dynamics365.UIAutomation.Sample.Web
                 xrmBrowser.Grid.OpenRecord(0);
 
                 //This uses the unique name (not label) from the subgrid properties
-                xrmBrowser.Grid.DeleteSubGridRecord("CONTACTS", 0);
+                xrmBrowser.Entity.DeleteSubGridRecord("CONTACTS", 0);
 
                 xrmBrowser.ThinkTime(2000);
             }
@@ -77,10 +77,10 @@ namespace Microsoft.Dynamics365.UIAutomation.Sample.Web
                 xrmBrowser.Grid.OpenRecord(0);
 
                 //This uses the unique name (not label) from the subgrid properties
-                List<GridItem> contacts = xrmBrowser.Grid.GetSubGridItems("CONTACTS");
+                List<GridItem> contacts = xrmBrowser.Entity.GetSubGridItems("CONTACTS");
 
                 //This uses the unique name (not label) from the subgrid properties
-                var contactCount = xrmBrowser.Grid.GetSubGridItemsCount("CONTACTS");
+                var contactCount = xrmBrowser.Entity.GetSubGridItemsCount("CONTACTS");
 
                 xrmBrowser.ThinkTime(2000);
             }
